+++ NEWS.md
@@ -10,6 +10,7 @@ Fixes:
 Enhancements:
   * Perform early quality checks on the content of 'sections' before advancing with migration and residency analysis.
   * Allow user to hide release sites from the study area diagrams.
+  * Improve spatial.csv vs spatial.txt mismatch error messaging by showing missing arrays.
 
 ## actel 0.1.3